+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0]
 ### Added
 - Added methods for start/end of line/input: `startOfLine()`, `startOfInput()`, `endOfLine()` and `endOfInput()`.
 - Added methods for a tab and line break: `tab()` and `lineBreak()`.
@@ -33,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - First setup with minimal features.
 
-[Unreleased]: https://github.com/ricoapon/readable-regex/compare/v0.2...HEAD
-[0.2.0]: https://github.com/ricoapon/readable-regex/releases/tag/v0.2
+[Unreleased]: https://github.com/ricoapon/readable-regex/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ricoapon/readable-regex/releases/tag/v0.3.0
+[0.2.0]: https://github.com/ricoapon/readable-regex/releases/tag/v0.2.0
 [0.1]: https://github.com/ricoapon/readable-regex/releases/tag/v0.1